Fixes for 3.0+
ImageSharp seems to be having trouble determining if Spans are available
and is loading the wrong assembly at runtime. Upgrading to the latest
version appears to fix this. This also changes a few things with how
the app renders pixels.

I think this means it also requires .NET Core 3.0+ but I need to do more
digging.

Closes #1